refactor: rename api files to reflect class name
