+++ src/api/pair-liquidity-info-history/pair-liquidity-info-history.controller.ts
@@ -5,7 +5,7 @@ import {
   ParseIntPipe,
   Query,
 } from '@nestjs/common';
-import { PairLiquidityInfoHistoryService } from './service';
+import { PairLiquidityInfoHistoryService } from './pair-liquidity-info-history.service';
 import { ApiOperation, ApiQuery, ApiResponse } from '@nestjs/swagger';
 import * as dto from '../../dto';
 import { OrderQueryEnum } from '../../dto';
