+++ src/api/pair-liquidity-info-history/pair-liquidity-info-history.module.ts
@@ -1,6 +1,6 @@
 import { Module } from '@nestjs/common';
-import { PairLiquidityInfoHistoryService } from './service';
-import { PairLiquidityInfoHistoryController } from './controller';
+import { PairLiquidityInfoHistoryService } from './pair-liquidity-info-history.service';
+import { PairLiquidityInfoHistoryController } from './pair-liquidity-info-history.controller';
 import { DatabaseModule } from '../../database/database.module';
 
 @Module({
