+++ src/api/pairs/pairs.service.ts
