+++ test/pairs.e2e-spec.ts
@@ -5,8 +5,8 @@ import { listToken, mockContext, sortByAddress } from './utils';
 import createWorkerMethods from '../src/worker';
 import * as data from './data/context-mockups';
 import * as db from './utils/db';
-import { TokensModule } from '../src/api/tokens/module';
-import { PairsModule } from '../src/api/pairs/module';
+import { TokensModule } from '../src/api/tokens/tokens.module';
+import { PairsModule } from '../src/api/pairs/pairs.module';
 
 type WorkerMethods = ReturnType<typeof createWorkerMethods>;
 let activeWorker: WorkerMethods;
