+++ test/swap-routes.e2e-spec.ts
@@ -6,8 +6,8 @@ import { listToken, mockContext } from './utils';
 import worker from '../src/worker';
 import * as db from './utils/db';
 import * as data from './data/context-mockups';
-import { TokensModule } from '../src/api/tokens/module';
-import { PairsModule } from '../src/api/pairs/module';
+import { TokensModule } from '../src/api/tokens/tokens.module';
+import { PairsModule } from '../src/api/pairs/pairs.module';
 
 type WorkerMethods = ReturnType<typeof worker>;
 let activeWorker: WorkerMethods;
